# api/openapi/RollForBrew.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RollForBrew.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "RollForBrew.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "RollForBrew.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/acting-as": {
      "get": {
        "tags": [
          "RollForBrew.Api"
        ],
        "operationId": "getActingAs",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActingAsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/brew-ratings/{roundId}": {
      "put": {
        "tags": [
          "RollForBrew.Api"
        ],
        "operationId": "submitBrewRating",
        "parameters": [
          {
            "name": "roundId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingIdResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RollForBrew.Api"
        ],
        "operationId": "withdrawBrewRating",
        "parameters": [
          {
            "name": "roundId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/brew-ratings/{roundId}/mine": {
      "get": {
        "tags": [
          "RollForBrew.Api"
        ],
        "operationId": "getMyBrewRating",
        "parameters": [
          {
            "name": "roundId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRatingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/spell-card-ratings/{cardId}": {
      "put": {
        "tags": [
          "RollForBrew.Api"
        ],
        "operationId": "rateSpellCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingIdResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RollForBrew.Api"
        ],
        "operationId": "withdrawSpellCardRating",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/spell-card-ratings/{cardId}/mine": {
      "get": {
        "tags": [
          "RollForBrew.Api"
        ],
        "operationId": "getMySpellCardRating",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRatingResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActingAsResponse": {
        "required": [
          "actingAsPlayerId"
        ],
        "type": "object",
        "properties": {
          "actingAsPlayerId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "MyRatingResponse": {
        "required": [
          "score"
        ],
        "type": "object",
        "properties": {
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "RatingIdResponse": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ScoreRequest": {
        "required": [
          "score"
        ],
        "type": "object",
        "properties": {
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "RollForBrew.Api"
    }
  ]
}